Updating redux init to default to prod.
Also fixing a bug where it wasn’t going into prod mode at all.
